% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texttable.R
\name{texttable}
\alias{texttable}
\title{Import tables from various text formats}
\usage{
texttable(text, format = NULL, trim_leading = TRUE, ...)
}
\arguments{
\item{text}{A file, URL, or character string. Only length 1 inputs are allowed.}

\item{format}{The format of the input text.  The default will guess from the
file extension of a file or URL, or default to Pandoc-flavored
markdown for input passed as a character string. For allowable
input types, see the
\href{http://pandoc.org/README.html#general-options}{Pandoc manual}.}

\item{trim_leading}{Should leading whitespace be trimmed from text input?}

\item{...}{arguments to pass to \code{\link[rvest]{html_table}} to specify how tables
should be parsed.}
}
\value{
A list of data frames.  Any tables with captions will have the
        caption as the list element name.
}
\description{
Parse tables from various formats into data frames.
}
\details{
\code{texttable} converts the input to HTML via
         \href{http://pandoc.org/}{Pandoc} and then imports tables via
         \code{\link[rvest]{html_table}}.

         Due to issues in pandoc, ODT and DocBook formats do not currently
         work.
}

